EntityBase & EF Core

+++ BoatRentalSystem.API/Controllers/AdditionController.cs
@@ -1,5 +1,7 @@
 ﻿namespace BoatRentalSystem.API.Controllers
 {
+    using AutoMapper;
+    using BoatRentalSystem.API.ViewModels;
     using BoatRentalSystem.Application;
     using BoatRentalSystem.Core.Entities;
     using Microsoft.AspNetCore.Http;
@@ -10,40 +12,65 @@
     public class AdditionController : ControllerBase
     {
         private readonly AdditionService _additionService;
+        private readonly IMapper _mapper;
 
-        public AdditionController(AdditionService additionService)
+        public AdditionController(AdditionService additionService, IMapper mapper)
         {
             _additionService = additionService;
+            _mapper = mapper;
         }
 
         [HttpGet]
-        public Task<IEnumerable<Addition>> Get()
+        public async Task<ActionResult<IEnumerable<AdditionViewModel>>> Get()
         {
-            return _additionService.GetAllAdditions();
+            var addition = await _additionService.GetAllAdditions();
+            var additionViewModel = _mapper.Map<IEnumerable<AdditionViewModel>>(addition);
+            return Ok(additionViewModel);
         }
 
         [HttpGet("{id}")]
-        public Task<Addition> Get(int id)
+        public async Task<ActionResult<AdditionViewModel>> Get(int id)
         {
-            return _additionService.GetAdditionById(id);
+            var addition = await _additionService.GetAdditionById(id);
+            if (addition == null)
+            {
+                return NotFound();
+            }
+            var additionViewModel = _mapper.Map<AdditionViewModel>(addition);
+            return Ok(additionViewModel);
         }
 
         [HttpPost]
-        public Task Post(Addition addition)
+        public async Task<ActionResult> Post([FromBody] AddAdditionViewModel addAdditionViewModel)
         {
-            return _additionService.AddAddition(addition);
+            var addition = _mapper.Map<Addition>(addAdditionViewModel);
+            await _additionService.AddAddition(addition);
+            return CreatedAtAction(nameof(Get), new { id = addition.Id }, addAdditionViewModel);
         }
 
         [HttpPut]
-        public Task Put(Addition addition)
+        public async Task<ActionResult> Put(AdditionViewModel additionViewModel)
         {
-            return _additionService.UpdateAddition(addition);
+            var existingAddition = await _additionService.GetAdditionById(additionViewModel.Id);
+            if (existingAddition == null)
+            {
+                return NotFound();
+            }
+            var addition = _mapper.Map<Addition>(additionViewModel);
+            await _additionService.UpdateAddition(addition);
+            return Ok(addition);
 
         }
         [HttpDelete]
-        public Task Delete(int id)
+        public async Task<ActionResult> Delete(int id)
         {
-            return _additionService.DeleteAddition(id);
+            var existingAddition = await _additionService.GetAdditionById(id);
+            if (existingAddition == null)
+            {
+                return NotFound();
+            }
+            await _additionService.DeleteAddition(id);
+            return NoContent();
         }
     }
 }
