+++ BoatRentalSystem.API/Controllers/CityController.cs
@@ -1,5 +1,7 @@
 ﻿namespace BoatRentalSystem.API.Controllers
 {
+    using AutoMapper;
+    using BoatRentalSystem.API.ViewModels;
     using BoatRentalSystem.Application;
     using BoatRentalSystem.Core.Entities;
     using Microsoft.AspNetCore.Http;
@@ -10,40 +12,65 @@
     public class CityController : ControllerBase
     {
         private readonly CityService _cityService;
+        private readonly IMapper _mapper;
 
-        public CityController(CityService cityService)
+        public CityController(CityService cityService, IMapper mapper)
         {
             _cityService = cityService;
+            _mapper = mapper;
         }
 
         [HttpGet]
-        public Task<IEnumerable<City>> Get()
+        public async Task<ActionResult<IEnumerable<CityViewModel>>> Get()
         {
-            return _cityService.GetAllCities();
+            var city = await _cityService.GetAllCities();
+            var cityViewModel = _mapper.Map<IEnumerable<CityViewModel>>(city);
+            return Ok(cityViewModel);
         }
 
         [HttpGet("{id}")]
-        public Task<City> Get(int id)
+        public async Task<ActionResult<CityViewModel>> Get(int id)
         {
-            return _cityService.GetCityById(id);
+            var city = await _cityService.GetCityById(id);
+            if (city == null)
+            {
+                return NotFound();
+            }
+            var cityViewModel = _mapper.Map<CityViewModel>(city);
+            return Ok(cityViewModel);
         }
 
         [HttpPost]
-        public Task Post(City city)
+        public async Task<ActionResult> Post([FromBody] AddCityViewModel addCityViewModel)
         {
-            return _cityService.AddCity(city);
+            var city = _mapper.Map<City>(addCityViewModel);
+            await _cityService.AddCity(city);
+            return CreatedAtAction(nameof(Get), new { id = city.Id }, addCityViewModel);
         }
 
         [HttpPut]
-        public Task Put(City city)
+        public async Task<ActionResult> Put(CityViewModel cityViewModel)
         {
-            return _cityService.UpdateCity(city);
+            var existingCity = await _cityService.GetCityById(cityViewModel.Id);
+            if (existingCity == null)
+            {
+                return NotFound();
+            }
+            var city = _mapper.Map<City>(cityViewModel);
+            await _cityService.UpdateCity(city);
+            return Ok(city);
 
         }
         [HttpDelete]
-        public Task Delete(int id)
+        public async Task<ActionResult> Delete(int id)
         {
-            return _cityService.DeleteCity(id);
+            var existingCity = await _cityService.GetCityById(id);
+            if (existingCity == null)
+            {
+                return NotFound();
+            }
+            await _cityService.DeleteCity(id);
+            return NoContent();
         }
     }
 }
